Svelte 5 TS Adapter + Svelte 5 playground (Up to date)
This is just a backup as I am currently reworking the main branch and Inertia is not planning to have a separate Svelte 5 adapter.

Note: There is currently a bug with NPM where packages are not properly scoped so you will get ParseError: Unexpected character '@' as an error. Using pnpm works as expected.

It is possible that with all other workspace packages up to date in this branch it resolves itself.